Fixed Data Import
Generates a SQL query command list to import an MBA data export in CSV merging with an existing table.  The tool can be used to generate a script to import a single file or in combination with other tools to parse an entire archive and import any arbitrary set of MBA data that is from a validated set used in annual reports or with raw monthly exports.

The script assumes an MBA schema that includes a numeric 'year' and boolean 'validated' column.  These columns are recommended (but not required) for managing mixed datasets that include both validated and raw MBA CSV data exports.

The script takes a single argument for the path and file to the target CSV, and an option "v" to indicate a validated data set.